Split the FutureProcessor into multiple helpers.
I split the FutureProcessor into helpers, one per kind of action
that it performs.

These helpers are mostly self contained and easy to test in isolation.

Every helper has docs about its side effects.
